refactor: move primitive-hooks into the primitive code files
