--- packages/react/src/primitive-hooks/actionBar/useActionBarFeedbackNegative.tsx
@@ -1,14 +1,0 @@
-"use client";
-
-import { useCallback } from "react";
-import { useMessageRuntime } from "../../context";
-
-export const useActionBarFeedbackNegative = () => {
-  const messageRuntime = useMessageRuntime();
-
-  const callback = useCallback(() => {
-    messageRuntime.submitFeedback({ type: "negative" });
-  }, [messageRuntime]);
-
-  return callback;
-};
