--- packages/react/src/primitive-hooks/actionBar/useActionBarFeedbackPositive.tsx
@@ -1,14 +1,0 @@
-"use client";
-
-import { useCallback } from "react";
-import { useMessageRuntime } from "../../context";
-
-export const useActionBarFeedbackPositive = () => {
-  const messageRuntime = useMessageRuntime();
-
-  const callback = useCallback(() => {
-    messageRuntime.submitFeedback({ type: "positive" });
-  }, [messageRuntime]);
-
-  return callback;
-};
