--- packages/react/src/primitive-hooks/branchPicker/useBranchPickerCount.tsx
@@ -1,8 +1,0 @@
-"use client";
-
-import { useMessage } from "../../context/react/MessageContext";
-
-export const useBranchPickerCount = () => {
-  const branchCount = useMessage((s) => s.branchCount);
-  return branchCount;
-};
