--- packages/react/src/primitive-hooks/branchPicker/useBranchPickerNext.tsx
@@ -1,19 +1,0 @@
-"use client";
-
-import { useCallback } from "react";
-import {
-  useMessage,
-  useMessageRuntime,
-} from "../../context/react/MessageContext";
-
-export const useBranchPickerNext = () => {
-  const messageRuntime = useMessageRuntime();
-  const disabled = useMessage((m) => m.branchNumber >= m.branchCount);
-
-  const callback = useCallback(() => {
-    messageRuntime.switchToBranch({ position: "next" });
-  }, [messageRuntime]);
-
-  if (disabled) return null;
-  return callback;
-};
