--- packages/react/src/primitive-hooks/branchPicker/useBranchPickerNumber.tsx
@@ -1,8 +1,0 @@
-"use client";
-
-import { useMessage } from "../../context/react/MessageContext";
-
-export const useBranchPickerNumber = () => {
-  const branchNumber = useMessage((s) => s.branchNumber);
-  return branchNumber;
-};
