--- packages/react/src/primitive-hooks/branchPicker/useBranchPickerPrevious.tsx
@@ -1,19 +1,0 @@
-"use client";
-
-import { useCallback } from "react";
-import {
-  useMessage,
-  useMessageRuntime,
-} from "../../context/react/MessageContext";
-
-export const useBranchPickerPrevious = () => {
-  const messageRuntime = useMessageRuntime();
-  const disabled = useMessage((m) => m.branchNumber <= 1);
-
-  const callback = useCallback(() => {
-    messageRuntime.switchToBranch({ position: "previous" });
-  }, [messageRuntime]);
-
-  if (disabled) return null;
-  return callback;
-};
