--- packages/react/src/primitive-hooks/contentPart/index.ts
@@ -1,3 +1,0 @@
-export { useContentPartDisplay } from "./useContentPartDisplay";
-export { useContentPartImage } from "./useContentPartImage";
-export { useContentPartText } from "./useContentPartText";
