--- packages/react/src/primitive-hooks/index.ts
@@ -1,1 +1,0 @@
-export * from "./contentPart";
