--- packages/react/src/primitive-hooks/thread/useThreadEmpty.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import { useThreadIf } from "./useThreadIf";
-
-export const useThreadEmpty = () => {
-  return useThreadIf({ empty: true });
-};
