+++ packages/react/src/primitives/actionBar/ActionBarEdit.tsx
@@ -1,11 +1,24 @@
 "use client";
 
-import { useActionBarEdit } from "../../primitive-hooks/actionBar/useActionBarEdit";
 import {
   ActionButtonElement,
   ActionButtonProps,
   createActionButton,
 } from "../../utils/createActionButton";
+import { useCallback } from "react";
+import { useEditComposer, useMessageRuntime } from "../../context";
+
+const useActionBarEdit = () => {
+  const messageRuntime = useMessageRuntime();
+  const disabled = useEditComposer((c) => c.isEditing);
+
+  const callback = useCallback(() => {
+    messageRuntime.composer.beginEdit();
+  }, [messageRuntime]);
+
+  if (disabled) return null;
+  return callback;
+};
 
 export namespace ActionBarPrimitiveEdit {
   export type Element = ActionButtonElement;
