+++ packages/react/src/primitives/composer/ComposerAddAttachment.tsx
@@ -5,7 +5,42 @@ import {
   ActionButtonProps,
   createActionButton,
 } from "../../utils/createActionButton";
-import { useComposerAddAttachment } from "../../primitive-hooks/composer/useComposerAddAttachment";
+import { useCallback } from "react";
+import { useComposer, useComposerRuntime } from "../../context";
+
+const useComposerAddAttachment = ({
+  multiple = true,
+}: {
+  /** allow selecting multiple files */
+  multiple?: boolean | undefined;
+} = {}) => {
+  const disabled = useComposer((c) => !c.isEditing);
+
+  const composerRuntime = useComposerRuntime();
+  const callback = useCallback(() => {
+    const input = document.createElement("input");
+    input.type = "file";
+    input.multiple = multiple;
+
+    const attachmentAccept = composerRuntime.getAttachmentAccept();
+    if (attachmentAccept !== "*") {
+      input.accept = attachmentAccept;
+    }
+
+    input.onchange = (e) => {
+      const fileList = (e.target as HTMLInputElement).files;
+      if (!fileList) return;
+      for (const file of fileList) {
+        composerRuntime.addAttachment(file);
+      }
+    };
+
+    input.click();
+  }, [composerRuntime]);
+
+  if (disabled) return null;
+  return callback;
+};
 
 export namespace ComposerPrimitiveAddAttachment {
   export type Element = ActionButtonElement;
