+++ packages/react/src/primitives/composer/ComposerCancel.tsx
@@ -5,7 +5,20 @@ import {
   ActionButtonProps,
   createActionButton,
 } from "../../utils/createActionButton";
-import { useComposerCancel } from "../../primitive-hooks/composer/useComposerCancel";
+import { useCallback } from "react";
+import { useComposer, useComposerRuntime } from "../../context";
+
+const useComposerCancel = () => {
+  const composerRuntime = useComposerRuntime();
+  const disabled = useComposer((c) => !c.canCancel);
+
+  const callback = useCallback(() => {
+    composerRuntime.cancel();
+  }, [composerRuntime]);
+
+  if (disabled) return null;
+  return callback;
+};
 
 export namespace ComposerPrimitiveCancel {
   export type Element = ActionButtonElement;
