+++ packages/react/src/primitives/contentPart/useContentPartImage.tsx
