+++ packages/react/src/primitives/index.ts
@@ -8,3 +8,7 @@ export * as MessagePrimitive from "./message";
 export * as ThreadPrimitive from "./thread";
 export * as ThreadListPrimitive from "./threadList";
 export * as ThreadListItemPrimitive from "./threadListItem";
+
+export { useContentPartDisplay } from "./contentPart/useContentPartDisplay";
+export { useContentPartImage } from "./contentPart/useContentPartImage";
+export { useContentPartText } from "./contentPart/useContentPartText";
