+++ packages/react/src/primitives/thread/ThreadIf.tsx
@@ -1,10 +1,29 @@
 "use client";
 
 import type { FC, PropsWithChildren } from "react";
-import {
-  UseThreadIfProps,
-  useThreadIf,
-} from "../../primitive-hooks/thread/useThreadIf";
+import { useThread } from "../../context";
+import type { RequireAtLeastOne } from "../../utils/RequireAtLeastOne";
+
+type ThreadIfFilters = {
+  empty: boolean | undefined;
+  running: boolean | undefined;
+  disabled: boolean | undefined;
+};
+
+type UseThreadIfProps = RequireAtLeastOne<ThreadIfFilters>;
+
+const useThreadIf = (props: UseThreadIfProps) => {
+  return useThread((thread) => {
+    if (props.empty === true && thread.messages.length !== 0) return false;
+    if (props.empty === false && thread.messages.length === 0) return false;
+    if (props.running === true && !thread.isRunning) return false;
+    if (props.running === false && thread.isRunning) return false;
+    if (props.disabled === true && thread.isDisabled) return false;
+    if (props.disabled === false && thread.isDisabled) return false;
+
+    return true;
+  });
+};
 
 export namespace ThreadPrimitiveIf {
   export type Props = PropsWithChildren<UseThreadIfProps>;
