+++ packages/react/src/primitives/thread/ThreadSuggestion.tsx
@@ -5,7 +5,33 @@ import {
   ActionButtonProps,
   createActionButton,
 } from "../../utils/createActionButton";
-import { useThreadSuggestion } from "../../primitive-hooks/thread/useThreadSuggestion";
+import { useCallback } from "react";
+import { useThread } from "../../context";
+import { useThreadRuntime } from "../../context/react/ThreadContext";
+
+const useThreadSuggestion = ({
+  prompt,
+  autoSend,
+}: {
+  prompt: string;
+  method: "replace";
+  autoSend?: boolean | undefined;
+}) => {
+  const threadRuntime = useThreadRuntime();
+
+  const disabled = useThread((t) => t.isDisabled);
+  const callback = useCallback(() => {
+    if (autoSend && !threadRuntime.getState().isRunning) {
+      threadRuntime.append(prompt);
+      threadRuntime.composer.setText("");
+    } else {
+      threadRuntime.composer.setText(prompt);
+    }
+  }, [threadRuntime, autoSend, prompt]);
+
+  if (disabled) return null;
+  return callback;
+};
 
 export namespace ThreadPrimitiveSuggestion {
   export type Element = ActionButtonElement;
